+ Use static_cast to eliminate a warning about comparing signed and   unsigned integers.

+++ src/c4/C4_MPI.cc
@@ -100,7 +100,7 @@ std::string processor_name()
     
     MPI_Get_processor_name( processor_name, &namelen );
     std::string procname( processor_name );
-    Check( namelen == procname.size() );
+    Check( static_cast<size_t>(namelen) == procname.size() );
     return procname;
 }
